+++ Engine/common_gl/source/GLIndexBuffer.cpp
@@ -5,11 +5,11 @@
 
 #include "glCheck.hpp"
 
-GLIndexBuffer::GLIndexBuffer(std::vector<uint16_t>* indices) : count(static_cast<GLsizei>(indices->size()))
+GLIndexBuffer::GLIndexBuffer(std::vector<uint32_t>* indices) : count(static_cast<GLsizei>(indices->size()))
 {
 	glCheck(glCreateBuffers(1, &indexHandle));
-	//glCheck(glNamedBufferStorage(indexHandle, sizeof(uint16_t) * count, indices->data(), GL_DYNAMIC_STORAGE_BIT));
-	glCheck(glNamedBufferData(indexHandle, sizeof(uint16_t) * count, indices->data(), GL_DYNAMIC_DRAW));
+	//glCheck(glNamedBufferStorage(indexHandle, sizeof(uint32_t) * count, indices->data(), GL_DYNAMIC_STORAGE_BIT));
+	glCheck(glNamedBufferData(indexHandle, sizeof(uint32_t) * count, indices->data(), GL_DYNAMIC_DRAW));
 }
 
 GLIndexBuffer::~GLIndexBuffer()
